FINALLY FIXED APPROXIMATE LAMBDA OPERATOR!
The boundary value at depth was screwing up. It turns out for
upward-pointing rays, the coefficient for the matrix element at depth
needs to be 1, not 0. (This is confusing to me, given that Olson &
Kunasz (1987) says to "set the incident boundary intensities to zero".)

I'll need to think a lot more about why this is the right answer. But at
least for now it works. So I should be able to strip out any bandwidth
of this matrix to get a correct Lambda_star.